move convert_vector to separate module
