+++ mapchete/io/vector/__init__.py
@@ -1,7 +1,7 @@
+from mapchete.io.vector.convert import convert_vector
+from mapchete.io.vector.indexed_features import IndexedFeatures
 from mapchete.io.vector.open import fiona_open
 from mapchete.io.vector.read import (
-    IndexedFeatures,
-    convert_vector,
     fiona_read,
     read_vector,
     read_vector_window,
